Settings tab overhaul
