+++ Anamnesis/Languages/en.json
@@ -744,6 +744,15 @@
   "Pose_iv_asi_kusu_a_r": "IVCS Fore Toe A Right",
   "Pose_iv_asi_kusu_b_r": "IVCS Fore Toe B Right",
 
+  "SettingsPages_General": "General",
+  "SettingsPages_General_Tooltip": "General application settings",
+  "SettingsPages_Personalization": "Personalization",
+  "SettingsPages_Personalization_Tooltip": "Application personalization",
+  "SettingsPages_Input": "Input & Hotkeys",
+  "SettingsPages_Input_Tooltip": "Input-related settings",
+  "SettingsPages_Connections": "Connections",
+  "SettingsPages_Connections_Tooltip": "Manage connections with external services",
+
   "Settings_Header": "Settings",
   "Settings_InterfaceHeader": "Interface",
   "Settings_FilesHeader": "Files",
@@ -781,7 +790,7 @@
   "Settings_ExternalHeader": "External Services",
   "Settings_UseExternalRefresh": "Use Penumbra Redraw",
   "Settings_UseExternalRefreshBrio": "Enable Brio Integration",
-  "Settings_UseExternalRefreshWarning": "Requires the HTTP API to be enabled in Penumbra/Brio. This option is found under the Advanced/Integrations section of the Settings tab of the plugins.",
+  "Settings_UseExternalRefreshWarning": "Requires HTTP API to be enabled in Penumbra/Brio. This option can be found under the Advanced/Integrations section of the plugins' settings.",
   "Settings_EnableNpcHack": "Enable NPC Faces on PC",
   "Settings_EnableNpcHackWarning": "This will break Penumbra Collections.",
   "Settings_GameHotkeysTooltip": "Enable the use of hotkeys even when FFXIV has focus while in GPose.\nThis will prevent FFXIV from recieving the used keys.",
