+++ Anamnesis/Tabs/Settings/ConnectionsSettingsPage.xaml
@@ -1,0 +1,63 @@
+﻿<UserControl x:Class="Anamnesis.Tabs.Settings.ConnectionsSettingsPage"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:XivToolsWPF="clr-namespace:XivToolsWpf.Controls;assembly=XivToolsWpf"
+             mc:Ignorable="d"
+             d:DesignHeight="450">
+	<GroupBox x:Name="ContentArea" Style="{StaticResource PanelGroupBox}">
+		<GroupBox.Header>
+			<XivToolsWPF:Header Key="SettingsPages_Connections" Icon="NetworkWired"/>
+		</GroupBox.Header>
+
+		<ScrollViewer VerticalScrollBarVisibility="Auto">
+			<Grid HorizontalAlignment="Center" Width="450">
+				<Grid.RowDefinitions>
+					<RowDefinition Height="Auto"/>
+				</Grid.RowDefinitions>
+
+				<!-- External Services -->
+				<GroupBox Grid.Row="0" Margin="0, 0, 0, 10">
+					<GroupBox.Header>
+						<XivToolsWPF:Header Key="Settings_ExternalHeader" Icon="Link"/>
+					</GroupBox.Header>
+
+					<Grid>
+						<Grid.ColumnDefinitions>
+							<ColumnDefinition Width="Auto"/>
+							<ColumnDefinition/>
+						</Grid.ColumnDefinitions>
+
+						<Grid.RowDefinitions>
+							<RowDefinition Height="Auto"/>
+							<RowDefinition Height="Auto"/>
+							<RowDefinition Height="Auto"/>
+							<RowDefinition Height="Auto"/>
+							<RowDefinition Height="Auto"/>
+						</Grid.RowDefinitions>
+
+						<XivToolsWPF:TextBlock Grid.Column="0" Grid.Row="0" Key="Settings_UseExternalRefreshBrio" Style="{StaticResource Label}"/>
+						<CheckBox IsChecked="{Binding SettingsService.Settings.UseExternalRefreshBrio}" Grid.Column="1" Grid.Row="0" Margin="6" x:Name="Settings_UseExternalRefreshBrio"/>
+						<XivToolsWPF:TextBlock Grid.Column="0" Grid.Row="1" Key="Settings_UseExternalRefresh" Style="{StaticResource Label}"/>
+						<CheckBox IsChecked="{Binding SettingsService.Settings.UseExternalRefresh}" Grid.Column="1" Grid.Row="1" Margin="6" x:Name="Settings_UseExternalRefresh"/>
+
+
+						<XivToolsWPF:InfoControl Grid.Row="2" Key="Settings_UseExternalRefreshWarning" Grid.ColumnSpan="2"/>
+
+						<XivToolsWPF:TextBlock Grid.Column="0" Grid.Row="3" Key="Settings_EnableNpcHack" Style="{StaticResource Label}"/>
+						<CheckBox Grid.Column="1" Grid.Row="3" Margin="6">
+							<CheckBox.Style>
+								<Style TargetType="CheckBox" BasedOn="{StaticResource MaterialDesignCheckBox}">
+									<Setter Property="IsChecked" Value="{Binding SettingsService.Settings.EnableNpcHack}"/>
+								</Style>
+							</CheckBox.Style>
+						</CheckBox>
+
+						<XivToolsWPF:InfoControl Grid.Row="4" Key="Settings_EnableNpcHackWarning" Grid.ColumnSpan="2"/>
+					</Grid>
+				</GroupBox>
+			</Grid>
+		</ScrollViewer>
+	</GroupBox>
+</UserControl>
