+++ Anamnesis/Tabs/Settings/PersonalizationSettingsPage.xaml
@@ -1,0 +1,95 @@
+﻿<UserControl x:Class="Anamnesis.Tabs.Settings.PersonalizationSettingsPage"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+			 xmlns:wpf="http://materialdesigninxaml.net/winfx/xaml/themes"
+             xmlns:XivToolsWPF="clr-namespace:XivToolsWpf.Controls;assembly=XivToolsWpf"
+             mc:Ignorable="d" 
+             d:DesignHeight="450">
+	<GroupBox x:Name="ContentArea" Style="{StaticResource PanelGroupBox}">
+		<GroupBox.Header>
+			<XivToolsWPF:Header Key="SettingsPages_Personalization" Icon="Palette"/>
+		</GroupBox.Header>
+
+		<ScrollViewer VerticalScrollBarVisibility="Auto">
+			<Grid HorizontalAlignment="Center" Width="450">
+				<Grid.RowDefinitions>
+					<RowDefinition Height="Auto"/>
+					<RowDefinition Height="Auto"/>
+				</Grid.RowDefinitions>
+
+				<!-- Theme -->
+				<GroupBox Grid.Row="0" Margin="0, 0, 0, 10">
+					<GroupBox.Header>
+						<XivToolsWPF:Header Key="Settings_Theme" Icon="PaintBrush"/>
+					</GroupBox.Header>
+
+					<Grid>
+						<Grid.ColumnDefinitions>
+							<ColumnDefinition Width="Auto"/>
+							<ColumnDefinition/>
+						</Grid.ColumnDefinitions>
+
+						<Grid.RowDefinitions>
+							<RowDefinition Height="Auto"/>
+							<RowDefinition Height="Auto"/>
+							<RowDefinition/>
+							<RowDefinition Height="Auto"/>
+						</Grid.RowDefinitions>
+
+						<XivToolsWPF:TextBlock Grid.Column="0" Grid.Row="0" Key="Settings_Theme_Override" Style="{StaticResource Label}"/>
+						<CheckBox IsChecked="{Binding SettingsService.Settings.OverrideSystemTheme}" Grid.Column="1" Grid.Row="0" Margin="6"/>
+
+						<XivToolsWPF:TextBlock Grid.Column="0" Grid.Row="1" Key="Settings_Theme_Light" Style="{StaticResource Label}"
+											   IsEnabled="{Binding SettingsService.Current.OverrideSystemTheme}"/>
+						<CheckBox IsChecked="{Binding SettingsService.Settings.ThemeLight}" Grid.Column="1" Grid.Row="1" Margin="6"
+								  IsEnabled="{Binding SettingsService.Current.OverrideSystemTheme}"/>
+
+						<XivToolsWPF:TextBlock Grid.Row="2" Grid.Column="0" Key="Settings_Theme_Color" Style="{StaticResource Label}"
+											   IsEnabled="{Binding SettingsService.Current.OverrideSystemTheme}"/>
+						<wpf:ColorPicker Color="{Binding SettingsService.Settings.ThemeTrimColor}" Grid.Row="2" Grid.Column="1"
+										 IsEnabled="{Binding SettingsService.Current.OverrideSystemTheme}"/>
+					</Grid>
+				</GroupBox>
+				
+				<!-- Gallery -->
+				<GroupBox Grid.Row="1">
+					<GroupBox.Header>
+						<XivToolsWPF:Header Key="Settings_GalleryHeader" Icon="Image"/>
+					</GroupBox.Header>
+
+					<Grid>
+						<Grid.ColumnDefinitions>
+							<ColumnDefinition Width="Auto"/>
+							<ColumnDefinition/>
+							<ColumnDefinition Width="Auto"/>
+						</Grid.ColumnDefinitions>
+
+						<Grid.RowDefinitions>
+							<RowDefinition/>
+							<RowDefinition/>
+						</Grid.RowDefinitions>
+
+						<XivToolsWPF:TextBlock Grid.Column="0" Grid.Row="0" Key="Settings_Gallery" Style="{StaticResource Label}"/>
+						<ComboBox Grid.Column="1" Grid.Row="0" Grid.ColumnSpan="2" Margin="1" x:Name="GalleryCombobox" SelectionChanged="OnGalleryChanged"
+								  HorizontalAlignment="Left" Width="200">
+							<ComboBoxItem>
+								<XivToolsWPF:TextBlock Key="Settings_Gallery_None"/>
+							</ComboBoxItem>
+							<ComboBoxItem>
+								<XivToolsWPF:TextBlock Key="Settings_Gallery_Curated"/>
+							</ComboBoxItem>
+							<ComboBoxItem>
+								<XivToolsWPF:TextBlock Key="Settings_Gallery_Local"/>
+							</ComboBoxItem>
+						</ComboBox>
+
+						<TextBox Grid.Row="1" Grid.Column="1" Margin="3, 0, 0, 0" Style="{StaticResource MaterialDesignTextBox}" Text="{Binding SettingsService.Settings.GalleryDirectory}" IsEnabled="False"/>
+						<Button Grid.Row="1" Grid.Column="2" Margin="6, 3, 0, 0" Style="{StaticResource TransparentButton}" Content="..." Click="OnBrowseGallery"/>
+					</Grid>
+				</GroupBox>
+			</Grid>
+		</ScrollViewer>
+	</GroupBox>
+</UserControl>
